test: merge provisioning test suites
